Adjusted documentation and placing of structs so they are visible in the cargo docs
